Separate querystore from querytype. Separate parameter map from query preparation
