+++ common/src/main/java/org/ldbcouncil/snb/impls/workloads/QueryStore.java
@@ -1,5 +1,17 @@
 package org.ldbcouncil.snb.impls.workloads;
-
+/**
+ * QueryStore.java
+ * 
+ * This class stores functions to query definition files and to retrieve the following used in operation handlers:
+ * - Query definition strings
+ * - Parameter map (Map<String, Object>), with as default String objects as values
+ * - Prepared queries (using string substitution)
+ * 
+ * Implementations can extend this class and override functions to change e.g.
+ * - ParameterPrefix ()
+ * - ParameterPostfix (file extension of query files)
+ * - Parameter map with different type stored than a string.
+ */
 import com.google.common.collect.ImmutableMap;
 import org.ldbcouncil.snb.driver.DbException;
 import org.ldbcouncil.snb.driver.workloads.interactive.*;
@@ -14,27 +26,39 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-/**
- * Superclass of query stores.
- * TODO: Move this class to the driver.
- *
- * Note: we deliberately do not use the {@link org.ldbcouncil.snb.driver.Operation#parameterMap()} method, because
- *
- * <ol>
- *   <li>some formats necessitate to handle ids differently from simple longs</li>
- *   <li>we would like to avoid any costs for type checking (e.g. using instanceof for each value)</li>
- * </ol>
- */
+
 public abstract class QueryStore {
 
+    /**
+     * Converter used for converting result objects. 
+     * @return
+     */
     protected Converter getConverter() { return new Converter(); }
 
+    /**
+     * Parameter prefix used in query definitions. Defaults to '$'
+     * @return
+     */
     protected String getParameterPrefix() { return "$"; }
 
+    /**
+     * The file extension for query files. 
+     * @return
+     */
     protected String getParameterPostfix() { return ""; }
 
+    /**
+     * Stores the loaded queries.
+     */
     protected Map<QueryType, String> queries = new HashMap<>();
 
+    /**
+     * Load a query file
+     * @param path Path to the file
+     * @param filename The name of the query file
+     * @return Loaded query definition
+     * @throws DbException
+     */
     protected String loadQueryFromFile(String path, String filename) throws DbException {
         final String filePath = path + File.separator + filename;
         try {
@@ -45,130 +69,102 @@ public abstract class QueryStore {
         }
     }
 
-    protected String prepare(QueryType queryType, Map<String, String> parameterSubstitutions) {
+    /**
+     * Prepares the parameterized query using the parameter map
+     * @param queryType Type of query to prepare (QueryType)
+     * @param parameterSubstitutions The parameter map containing the subsitute values
+     * @return Prepared query string
+     */
+    protected String prepare(QueryType queryType, Map<String, Object> parameterSubstitutions) {
         String querySpecification = queries.get(queryType);
         for (String parameter : parameterSubstitutions.keySet()) {
             querySpecification = querySpecification.replace(
                     getParameterPrefix() + parameter + getParameterPostfix(),
-                    parameterSubstitutions.get(parameter)
+                    (String) parameterSubstitutions.get(parameter)
             );
         }
         return querySpecification;
     }
 
+    /**
+     * 
+     * @param queryType
+     * @return
+     */
     public String getParameterizedQuery(QueryType queryType) {
         return queries.get(queryType);
     }
 
-    public enum QueryType {
-
-        // interactive complex queries
-        InteractiveComplexQuery1 ("interactive-complex-1" ),
-        InteractiveComplexQuery2 ("interactive-complex-2" ),
-        InteractiveComplexQuery3 ("interactive-complex-3" ),
-        InteractiveComplexQuery4 ("interactive-complex-4" ),
-        InteractiveComplexQuery5 ("interactive-complex-5" ),
-        InteractiveComplexQuery6 ("interactive-complex-6" ),
-        InteractiveComplexQuery7 ("interactive-complex-7" ),
-        InteractiveComplexQuery8 ("interactive-complex-8" ),
-        InteractiveComplexQuery9 ("interactive-complex-9" ),
-        InteractiveComplexQuery10("interactive-complex-10"),
-        InteractiveComplexQuery11("interactive-complex-11"),
-        InteractiveComplexQuery12("interactive-complex-12"),
-        InteractiveComplexQuery13("interactive-complex-13"),
-        InteractiveComplexQuery14("interactive-complex-14"),
-        InteractiveComplexQuery3DurationAsFunction ("interactive-complex-3-duration-as-function" ),
-        InteractiveComplexQuery4DurationAsFunction ("interactive-complex-4-duration-as-function" ),
-        InteractiveComplexQuery7WithSecond("interactive-complex-7-with-second"),
-
-        // interactive short queries
-        InteractiveShortQuery1("interactive-short-1"),
-        InteractiveShortQuery2("interactive-short-2"),
-        InteractiveShortQuery3("interactive-short-3"),
-        InteractiveShortQuery4("interactive-short-4"),
-        InteractiveShortQuery5("interactive-short-5"),
-        InteractiveShortQuery6("interactive-short-6"),
-        InteractiveShortQuery7("interactive-short-7"),
-
-        // interactive updates (single queries)
-        InteractiveUpdate1("interactive-update-1"),
-        InteractiveUpdate2("interactive-update-2"),
-        InteractiveUpdate3("interactive-update-3"),
-        InteractiveUpdate4("interactive-update-4"),
-        InteractiveUpdate5("interactive-update-5"),
-        InteractiveUpdate6("interactive-update-6"),
-        InteractiveUpdate7("interactive-update-7"),
-        InteractiveUpdate8("interactive-update-8"),
-
-        // interactive updates (additional queries for systems that perform them as multiple queries)
-        InteractiveUpdate1AddPerson            ("interactive-update-1-add-person"),
-        InteractiveUpdate1AddPersonCompanies   ("interactive-update-1-add-person-companies"),
-        InteractiveUpdate1AddPersonEmails      ("interactive-update-1-add-person-emails"),
-        InteractiveUpdate1AddPersonLanguages   ("interactive-update-1-add-person-languages"),
-        InteractiveUpdate1AddPersonTags        ("interactive-update-1-add-person-tags"),
-        InteractiveUpdate1AddPersonUniversities("interactive-update-1-add-person-universities"),
-
-        InteractiveUpdate4AddForum             ("interactive-update-4-add-forum"),
-        InteractiveUpdate4AddForumTags         ("interactive-update-4-add-forum-tags"),
-
-        InteractiveUpdate6AddPost              ("interactive-update-6-add-post"),
-        InteractiveUpdate6AddPostTags          ("interactive-update-6-add-post-tags"),
-
-        InteractiveUpdate7AddComment           ("interactive-update-7-add-comment"),
-        InteractiveUpdate7AddCommentTags       ("interactive-update-7-add-comment-tags"),
-
-        // interactive updates (additional queries for system that insert content/imageFile using separate operations)
-        InteractiveUpdate6AddPostContent       ("interactive-update-6-add-post-content"),
-        InteractiveUpdate6AddPostImageFile     ("interactive-update-6-add-post-imagefile"),
-        ;
-
-        private String name;
-
-        QueryType(String name) {
-            this.name = name;
-        }
-
-        public String getName() {
-            return name;
-        }
-    }
-
+    /**
+     * Create QueryStore
+     * @param path: Path to the directory containing query files
+     * @param postfix The extension of the query files
+     * @throws DbException
+     */
     public QueryStore(String path, String postfix) throws DbException {
         for (QueryType queryType : QueryType.values()) {
             queries.put(queryType, loadQueryFromFile(path, queryType.getName() + postfix));
         }
     }
 
-    // Interactive complex reads
-
+    /**
+     * Get prepared Query1 string
+     * @param operation LdbcQuery1 operation containing parameter values
+     * @return Prepared Query1 string
+     */
     public String getQuery1(LdbcQuery1 operation) {
         return prepare(QueryType.InteractiveComplexQuery1, getQuery1Map(operation));
     }
 
-    public Map<String, String> getQuery1Map(LdbcQuery1 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query1 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery1 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery1Map(LdbcQuery1 operation) {
+        return new ImmutableMap.Builder<String, Object>()
                 .put(LdbcQuery1.PERSON_ID, getConverter().convertId(operation.getPersonIdQ1()))
                 .put(LdbcQuery1.FIRST_NAME, getConverter().convertString(operation.getFirstName()))
                 .build();
     }
 
+    /**
+     * Get prepared Query2 string
+     * @param operation LdbcQuery2 operation containing parameter values
+     * @return Prepared Query2 string
+     */
     public String getQuery2(LdbcQuery2 operation) {
         return prepare(QueryType.InteractiveComplexQuery2, getQuery2Map(operation));
     }
 
-    public Map<String, String> getQuery2Map(LdbcQuery2 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query2 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery2 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery2Map(LdbcQuery2 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery2.PERSON_ID, getConverter().convertId(operation.getPersonIdQ2()))
         .put(LdbcQuery2.MAX_DATE, getConverter().convertDate(operation.getMaxDate()))
         .build();
     }
 
+    /**
+     * Get prepared Query3 string
+     * @param operation LdbcQuery3 operation containing parameter values
+     * @return Prepared Query3 string
+     */
     public String getQuery3(LdbcQuery3 operation) {
         return prepare(QueryType.InteractiveComplexQuery3, getQuery3Map(operation));
     }
 
-    public Map<String, String> getQuery3Map(LdbcQuery3 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query3 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery3 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery3Map(LdbcQuery3 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery3.PERSON_ID, getConverter().convertId(operation.getPersonIdQ3()))
         .put(LdbcQuery3.COUNTRY_X_NAME, getConverter().convertString(operation.getCountryXName()))
         .put(LdbcQuery3.COUNTRY_Y_NAME, getConverter().convertString(operation.getCountryYName()))
@@ -177,329 +173,615 @@ public abstract class QueryStore {
         .build();
     }
 
+    /**
+     * Get prepared Query4 string
+     * @param operation LdbcQuery4 operation containing parameter values
+     * @return Prepared Query4 string
+     */
     public String getQuery4(LdbcQuery4 operation) {
         return prepare(QueryType.InteractiveComplexQuery4, getQuery4Map(operation));
     }
 
-    public Map<String, String> getQuery4Map(LdbcQuery4 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query4 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery4 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery4Map(LdbcQuery4 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery4.PERSON_ID, getConverter().convertId(operation.getPersonIdQ4()))
         .put(LdbcQuery4.START_DATE, getConverter().convertDate(operation.getStartDate()))
         .put(LdbcQuery4.DURATION_DAYS, getConverter().convertInteger(operation.getDurationDays()))
         .build();
     }
 
+    /**
+     * Get prepared Query5 string
+     * @param operation LdbcQuery5 operation containing parameter values
+     * @return Prepared Query5 string
+     */
     public String getQuery5(LdbcQuery5 operation) {
         return prepare(QueryType.InteractiveComplexQuery5, getQuery5Map(operation));
     }
 
-    public Map<String, String> getQuery5Map(LdbcQuery5 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query5 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery5 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery5Map(LdbcQuery5 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery5.PERSON_ID, getConverter().convertId(operation.getPersonIdQ5()))
         .put(LdbcQuery5.MIN_DATE, getConverter().convertDate(operation.getMinDate()))
         .build();
     }
 
+    /**
+     * Get prepared Query6 string
+     * @param operation LdbcQuery6 operation containing parameter values
+     * @return Prepared Query6 string
+     */
     public String getQuery6(LdbcQuery6 operation) {
         return prepare(QueryType.InteractiveComplexQuery6, getQuery6Map(operation));
     }
 
-    public Map<String, String> getQuery6Map(LdbcQuery6 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query6 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery6 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery6Map(LdbcQuery6 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery6.PERSON_ID, getConverter().convertId(operation.getPersonIdQ6()))
         .put(LdbcQuery6.TAG_NAME, getConverter().convertString(operation.getTagName()))
         .build();
     }
 
+    /**
+     * Get prepared Query7 string
+     * @param operation LdbcQuery7 operation containing parameter values
+     * @return Prepared Query7 string
+     */
     public String getQuery7(LdbcQuery7 operation) {
         return prepare(QueryType.InteractiveComplexQuery7, getQuery7Map(operation));
     }
-    public Map<String, String> getQuery7Map(LdbcQuery7 operation) {
-        return new ImmutableMap.Builder<String, String>()
+
+    /**
+     * Get Query7 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery7 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery7Map(LdbcQuery7 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery7.PERSON_ID, getConverter().convertId(operation.getPersonIdQ7()))
         .build();
     }
 
+    /**
+     * Get prepared Query8 string
+     * @param operation LdbcQuery8 operation containing parameter values
+     * @return Prepared Query8 string
+     */
     public String getQuery8(LdbcQuery8 operation) {
         return prepare(QueryType.InteractiveComplexQuery8, getQuery8Map(operation));
     }
 
-    public Map<String, String> getQuery8Map(LdbcQuery8 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query8 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery8 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery8Map(LdbcQuery8 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery8.PERSON_ID, getConverter().convertId(operation.getPersonIdQ8()))
         .build();
     }
 
+    /**
+     * Get prepared Query9 string
+     * @param operation LdbcQuery9 operation containing parameter values
+     * @return Prepared Query9 string
+     */
     public String getQuery9(LdbcQuery9 operation) {
         return prepare(QueryType.InteractiveComplexQuery9, getQuery9Map(operation));
     }
 
-    public Map<String, String> getQuery9Map(LdbcQuery9 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query9 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery9 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery9Map(LdbcQuery9 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery9.PERSON_ID, getConverter().convertId(operation.getPersonIdQ9()))
         .put(LdbcQuery9.MAX_DATE, getConverter().convertDate(operation.getMaxDate()))
         .build();
     }
 
+    /**
+     * Get prepared Query10 string
+     * @param operation LdbcQuery10 operation containing parameter values
+     * @return Prepared Query10 string
+     */
     public String getQuery10(LdbcQuery10 operation) {
         return prepare(QueryType.InteractiveComplexQuery10, getQuery10Map(operation));
     }
 
-    public Map<String, String> getQuery10Map(LdbcQuery10 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query10 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery10 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery10Map(LdbcQuery10 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery10.PERSON_ID, getConverter().convertId(operation.getPersonIdQ10()))
         .put(LdbcQuery10.MONTH, getConverter().convertInteger(operation.getMonth()))
         .build();
     }
 
+    /**
+     * Get prepared Query11 string
+     * @param operation LdbcQuery11 operation containing parameter values
+     * @return Prepared Query11 string
+     */
     public String getQuery11(LdbcQuery11 operation) {
         return prepare(QueryType.InteractiveComplexQuery11, getQuery11Map(operation));
     }
 
-    public Map<String, String> getQuery11Map(LdbcQuery11 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query11 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery11 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery11Map(LdbcQuery11 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery11.PERSON_ID, getConverter().convertId(operation.getPersonIdQ11()))
         .put(LdbcQuery11.COUNTRY_NAME, getConverter().convertString(operation.getCountryName()))
         .put(LdbcQuery11.WORK_FROM_YEAR, getConverter().convertInteger(operation.getWorkFromYear()))
         .build();
     }
 
+    /**
+     * Get prepared Query12 string
+     * @param operation LdbcQuery12 operation containing parameter values
+     * @return Prepared Query12 string
+     */
     public String getQuery12(LdbcQuery12 operation) {
         return prepare(QueryType.InteractiveComplexQuery12, getQuery12Map(operation));
     }
 
-    public Map<String, String> getQuery12Map(LdbcQuery12 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query12 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery12 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery12Map(LdbcQuery12 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery12.PERSON_ID, getConverter().convertId(operation.getPersonIdQ12()))
         .put(LdbcQuery12.TAG_CLASS_NAME, getConverter().convertString(operation.getTagClassName()))
         .build();
     }
 
+    /**
+     * Get prepared Query13 string
+     * @param operation LdbcQuery13 operation containing parameter values
+     * @return Prepared Query13 string
+     */
     public String getQuery13(LdbcQuery13 operation) {
         return prepare(QueryType.InteractiveComplexQuery13, getQuery13Map(operation));
     }
 
-    public Map<String,String> getQuery13Map (LdbcQuery13 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query13 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery13 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery13Map (LdbcQuery13 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery13.PERSON1_ID, getConverter().convertId(operation.getPerson1IdQ13StartNode()))
         .put(LdbcQuery13.PERSON2_ID, getConverter().convertId(operation.getPerson2IdQ13EndNode()))
         .build();
     }
 
+    /**
+     * Get prepared Query14 string
+     * @param operation LdbcQuery14 operation containing parameter values
+     * @return Prepared Query14 string
+     */
     public String getQuery14(LdbcQuery14 operation) {
         return prepare(QueryType.InteractiveComplexQuery14, getQuery14Map(operation));
     }
 
-    public Map<String, String> getQuery14Map (LdbcQuery14 operation) {
-        return new ImmutableMap.Builder<String, String>()
+    /**
+     * Get Query14 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcQuery14 operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getQuery14Map (LdbcQuery14 operation) {
+        return new ImmutableMap.Builder<String, Object>()
         .put(LdbcQuery14.PERSON1_ID, getConverter().convertId(operation.getPerson1IdQ14StartNode()))
         .put(LdbcQuery14.PERSON2_ID, getConverter().convertId(operation.getPerson2IdQ14EndNode()))
         .build();
     }
 
-    // Interactive short reads
-
+    /**
+     * Get prepared LdbcShortQuery1PersonProfile string
+     * @param operation LdbcShortQuery1PersonProfile operation containing parameter values
+     * @return Prepared LdbcShortQuery1PersonProfile string
+     */
     public String getShortQuery1PersonProfile(LdbcShortQuery1PersonProfile operation) {
         return prepare(
                 QueryType.InteractiveShortQuery1,getShortQuery1PersonProfileMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery1PersonProfileMap(LdbcShortQuery1PersonProfile operation) {
+    /**
+     * Get LdbcShortQuery1PersonProfile Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery1PersonProfile operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery1PersonProfileMap(LdbcShortQuery1PersonProfile operation) {
         return ImmutableMap.of(LdbcShortQuery1PersonProfile.PERSON_ID, getConverter().convertId(operation.getPersonIdSQ1()));
-        
     }
 
+    /**
+     * Get prepared LdbcShortQuery2PersonPosts string
+     * @param operation LdbcShortQuery2PersonPosts operation containing parameter values
+     * @return Prepared LdbcShortQuery2PersonPosts string
+     */
     public String getShortQuery2PersonPosts(LdbcShortQuery2PersonPosts operation) {
         return prepare(
                 QueryType.InteractiveShortQuery2, getShortQuery2PersonPostsMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery2PersonPostsMap(LdbcShortQuery2PersonPosts operation) {
+    /**
+     * Get LdbcShortQuery2PersonPosts Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery2PersonPosts operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery2PersonPostsMap(LdbcShortQuery2PersonPosts operation) {
         return ImmutableMap.of(LdbcShortQuery2PersonPosts.PERSON_ID, getConverter().convertId(operation.getPersonIdSQ2()));
     }
 
+    /**
+     * Get prepared LdbcShortQuery3PersonFriends string
+     * @param operation LdbcShortQuery3PersonFriends operation containing parameter values
+     * @return Prepared LdbcShortQuery3PersonFriends string
+     */
     public String getShortQuery3PersonFriends(LdbcShortQuery3PersonFriends operation) {
         return prepare(
                 QueryType.InteractiveShortQuery3, getShortQuery3PersonFriendsMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery3PersonFriendsMap(LdbcShortQuery3PersonFriends operation) {
+    /**
+     * Get LdbcShortQuery3PersonFriends Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery3PersonFriends operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery3PersonFriendsMap(LdbcShortQuery3PersonFriends operation) {
         return ImmutableMap.of(LdbcShortQuery3PersonFriends.PERSON_ID, getConverter().convertId(operation.getPersonIdSQ3()));
     }
 
+    /**
+     * Get prepared LdbcShortQuery4MessageContent string
+     * @param operation LdbcShortQuery4MessageContent operation containing parameter values
+     * @return Prepared LdbcShortQuery4MessageContent string
+     */
     public String getShortQuery4MessageContent(LdbcShortQuery4MessageContent operation) {
         return prepare(
                 QueryType.InteractiveShortQuery4, getShortQuery4MessageContentMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery4MessageContentMap(LdbcShortQuery4MessageContent operation) {
+    /**
+     * Get LdbcShortQuery4MessageContent Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery4MessageContent operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery4MessageContentMap(LdbcShortQuery4MessageContent operation) {
         return ImmutableMap.of(LdbcShortQuery4MessageContent.MESSAGE_ID, getConverter().convertId(operation.getMessageIdContent()));
     }
 
+    /**
+     * Get prepared LdbcShortQuery5MessageCreator string
+     * @param operation LdbcShortQuery5MessageCreator operation containing parameter values
+     * @return Prepared LdbcShortQuery5MessageCreator string
+     */
     public String getShortQuery5MessageCreator(LdbcShortQuery5MessageCreator operation) {
         return prepare(
                 QueryType.InteractiveShortQuery5, getShortQuery5MessageCreatorMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery5MessageCreatorMap(LdbcShortQuery5MessageCreator operation) {
+    /**
+     * Get LdbcShortQuery5MessageCreator Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery5MessageCreator operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery5MessageCreatorMap(LdbcShortQuery5MessageCreator operation) {
         return ImmutableMap.of(LdbcShortQuery5MessageCreator.MESSAGE_ID, getConverter().convertId(operation.getMessageIdCreator()));
     }
 
+    /**
+     * Get prepared LdbcShortQuery6MessageForum string
+     * @param operation LdbcShortQuery6MessageForum operation containing parameter values
+     * @return Prepared LdbcShortQuery6MessageForum string
+     */
     public String getShortQuery6MessageForum(LdbcShortQuery6MessageForum operation) {
         return prepare(
                 QueryType.InteractiveShortQuery6, getShortQuery6MessageForumMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery6MessageForumMap(LdbcShortQuery6MessageForum operation) {
+    /**
+     * Get LdbcShortQuery6MessageForum Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery6MessageForum operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery6MessageForumMap(LdbcShortQuery6MessageForum operation) {
         return ImmutableMap.of(LdbcShortQuery6MessageForum.MESSAGE_ID, getConverter().convertId(operation.getMessageForumId()));
     }
 
+    /**
+     * Get prepared LdbcShortQuery7MessageReplies string
+     * @param operation LdbcShortQuery7MessageReplies operation containing parameter values
+     * @return Prepared LdbcShortQuery7MessageReplies string
+     */
     public String getShortQuery7MessageReplies(LdbcShortQuery7MessageReplies operation) {
         return prepare(
                 QueryType.InteractiveShortQuery7, getShortQuery7MessageRepliesMap(operation)
         );
     }
 
-    public Map<String, String> getShortQuery7MessageRepliesMap(LdbcShortQuery7MessageReplies operation) {
+    /**
+     * Get LdbcShortQuery7MessageReplies Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcShortQuery7MessageReplies operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getShortQuery7MessageRepliesMap(LdbcShortQuery7MessageReplies operation) {
         return ImmutableMap.of(LdbcShortQuery7MessageReplies.MESSAGE_ID, getConverter().convertId(operation.getMessageRepliesId()));
     }
 
-    // Interactive updates
-
-    // Interactive updates, formulated as a single query each
-
+    
+    /**
+     * Get prepared UpdateQuery1 string
+     * This is used for single query updates.
+     * @param operation UpdateQuery1 operation containing parameter values
+     * @return Prepared UpdateQuery1 string
+     */
     public String getUpdate1Single(LdbcUpdate1AddPerson operation) {
         return prepare(
-                QueryType.InteractiveUpdate1,
-                new ImmutableMap.Builder<String, String>()
-                        .put(LdbcUpdate1AddPerson.PERSON_ID, getConverter().convertIdForInsertion(operation.getPersonId()))
-                        .put(LdbcUpdate1AddPerson.PERSON_FIRST_NAME, getConverter().convertString(operation.getPersonFirstName()))
-                        .put(LdbcUpdate1AddPerson.PERSON_LAST_NAME, getConverter().convertString(operation.getPersonLastName()))
-                        .put(LdbcUpdate1AddPerson.GENDER, getConverter().convertString(operation.getGender()))
-                        .put(LdbcUpdate1AddPerson.BIRTHDAY, getConverter().convertDate(operation.getBirthday()))
-                        .put(LdbcUpdate1AddPerson.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
-                        .put(LdbcUpdate1AddPerson.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
-                        .put(LdbcUpdate1AddPerson.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
-                        .put(LdbcUpdate1AddPerson.CITY_ID, getConverter().convertId(operation.getCityId()))
-                        .put(LdbcUpdate1AddPerson.WORK_AT, getConverter().convertOrganisations(operation.getWorkAt()))
-                        .put(LdbcUpdate1AddPerson.STUDY_AT, getConverter().convertOrganisations(operation.getStudyAt()))
-                        .put(LdbcUpdate1AddPerson.EMAILS, getConverter().convertStringList(operation.getEmails()))
-                        .put(LdbcUpdate1AddPerson.LANGUAGES, getConverter().convertStringList(operation.getLanguages()))
-                        .put(LdbcUpdate1AddPerson.TAG_IDS, getConverter().convertLongList(operation.getTagIds()))
-                        .build()
+                QueryType.InteractiveUpdate1, getUpdate1SingleMap(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery1 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate1AddPerson operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate1SingleMap(LdbcUpdate1AddPerson operation) {
+        return new ImmutableMap.Builder<String, Object>()
+        .put(LdbcUpdate1AddPerson.PERSON_ID, getConverter().convertIdForInsertion(operation.getPersonId()))
+        .put(LdbcUpdate1AddPerson.PERSON_FIRST_NAME, getConverter().convertString(operation.getPersonFirstName()))
+        .put(LdbcUpdate1AddPerson.PERSON_LAST_NAME, getConverter().convertString(operation.getPersonLastName()))
+        .put(LdbcUpdate1AddPerson.GENDER, getConverter().convertString(operation.getGender()))
+        .put(LdbcUpdate1AddPerson.BIRTHDAY, getConverter().convertDate(operation.getBirthday()))
+        .put(LdbcUpdate1AddPerson.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
+        .put(LdbcUpdate1AddPerson.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
+        .put(LdbcUpdate1AddPerson.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
+        .put(LdbcUpdate1AddPerson.CITY_ID, getConverter().convertId(operation.getCityId()))
+        .put(LdbcUpdate1AddPerson.WORK_AT, getConverter().convertOrganisations(operation.getWorkAt()))
+        .put(LdbcUpdate1AddPerson.STUDY_AT, getConverter().convertOrganisations(operation.getStudyAt()))
+        .put(LdbcUpdate1AddPerson.EMAILS, getConverter().convertStringList(operation.getEmails()))
+        .put(LdbcUpdate1AddPerson.LANGUAGES, getConverter().convertStringList(operation.getLanguages()))
+        .put(LdbcUpdate1AddPerson.TAG_IDS, getConverter().convertLongList(operation.getTagIds()))
+        .build();
+    }
+
+    /**
+     * Get prepared UpdateQuery2 string
+     * @param operation LdbcUpdate2AddPostLike operation containing parameter values
+     * @return Prepared UpdateQuery2 string
+     */
     public String getUpdate2(LdbcUpdate2AddPostLike operation) {
         return prepare(
-                QueryType.InteractiveUpdate2,
-                ImmutableMap.of(
-                        LdbcUpdate2AddPostLike.PERSON_ID, getConverter().convertId(operation.getPersonId()),
-                        LdbcUpdate2AddPostLike.POST_ID, getConverter().convertId(operation.getPostId()),
-                        LdbcUpdate2AddPostLike.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate())
-                )
+                QueryType.InteractiveUpdate2, getUpdate2Map(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery2 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate2AddPostLike operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate2Map(LdbcUpdate2AddPostLike operation) {
+        return ImmutableMap.of(
+            LdbcUpdate2AddPostLike.PERSON_ID, getConverter().convertId(operation.getPersonId()),
+            LdbcUpdate2AddPostLike.POST_ID, getConverter().convertId(operation.getPostId()),
+            LdbcUpdate2AddPostLike.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate())
+        );
+    }
+
+    /**
+     * Get prepared UpdateQuery3 string
+     * @param operation LdbcUpdate3AddCommentLike operation containing parameter values
+     * @return Prepared UpdateQuery3 string
+     */
     public String getUpdate3(LdbcUpdate3AddCommentLike operation) {
         return prepare(
-                QueryType.InteractiveUpdate3,
-                ImmutableMap.of(
-                        LdbcUpdate3AddCommentLike.PERSON_ID, getConverter().convertId(operation.getPersonId()),
-                        LdbcUpdate3AddCommentLike.COMMENT_ID, getConverter().convertId(operation.getCommentId()),
-                        LdbcUpdate3AddCommentLike.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate())
-                )
+                QueryType.InteractiveUpdate3, getUpdate3Map(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery3 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate3AddCommentLike operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate3Map(LdbcUpdate3AddCommentLike operation) {
+        return ImmutableMap.of(
+            LdbcUpdate3AddCommentLike.PERSON_ID, getConverter().convertId(operation.getPersonId()),
+            LdbcUpdate3AddCommentLike.COMMENT_ID, getConverter().convertId(operation.getCommentId()),
+            LdbcUpdate3AddCommentLike.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate())
+        );
+    }
+
+    /**
+     * Get prepared UpdateQuery4 string
+     * This is used for single query updates.
+     * @param operation LdbcUpdate4AddForum operation containing parameter values
+     * @return Prepared UpdateQuery4 string
+     */
     public String getUpdate4Single(LdbcUpdate4AddForum operation) {
         return prepare(
-                QueryType.InteractiveUpdate4,
-                ImmutableMap.of(
-                        LdbcUpdate4AddForum.FORUM_ID, getConverter().convertIdForInsertion(operation.getForumId()),
-                        LdbcUpdate4AddForum.FORUM_TITLE, getConverter().convertString(operation.getForumTitle()),
-                        LdbcUpdate4AddForum.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()),
-                        LdbcUpdate4AddForum.MODERATOR_PERSON_ID, getConverter().convertId(operation.getModeratorPersonId()),
-                        LdbcUpdate4AddForum.TAG_IDS, getConverter().convertLongList(operation.getTagIds())
-                )
+                QueryType.InteractiveUpdate4, getUpdate4SingleMap(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery4 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate4AddForum operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate4SingleMap(LdbcUpdate4AddForum operation) {
+        return ImmutableMap.of(
+            LdbcUpdate4AddForum.FORUM_ID, getConverter().convertIdForInsertion(operation.getForumId()),
+            LdbcUpdate4AddForum.FORUM_TITLE, getConverter().convertString(operation.getForumTitle()),
+            LdbcUpdate4AddForum.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()),
+            LdbcUpdate4AddForum.MODERATOR_PERSON_ID, getConverter().convertId(operation.getModeratorPersonId()),
+            LdbcUpdate4AddForum.TAG_IDS, getConverter().convertLongList(operation.getTagIds())
+        );
+    }
 
+    /**
+     * Get prepared UpdateQuery5 string
+     * @param operation LdbcUpdate5AddForumMembership operation containing parameter values
+     * @return Prepared UpdateQuery5 string
+     */
     public String getUpdate5(LdbcUpdate5AddForumMembership operation) {
         return prepare(
-                QueryType.InteractiveUpdate5,
-                ImmutableMap.of(
-                        LdbcUpdate5AddForumMembership.FORUM_ID, getConverter().convertId(operation.getForumId()),
-                        LdbcUpdate5AddForumMembership.PERSON_ID, getConverter().convertId(operation.getPersonId()),
-                        LdbcUpdate5AddForumMembership.JOIN_DATE, getConverter().convertDateTime(operation.getJoinDate())
-                )
+                QueryType.InteractiveUpdate5, getUpdate5Map(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery5 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate5AddForumMembership operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate5Map(LdbcUpdate5AddForumMembership operation) {
+        return ImmutableMap.of(
+            LdbcUpdate5AddForumMembership.FORUM_ID, getConverter().convertId(operation.getForumId()),
+            LdbcUpdate5AddForumMembership.PERSON_ID, getConverter().convertId(operation.getPersonId()),
+            LdbcUpdate5AddForumMembership.JOIN_DATE, getConverter().convertDateTime(operation.getJoinDate())
+        );
+    }
+    
+    /**
+     * Get prepared UpdateQuery6 string
+     * This is used for single query updates.
+     * @param operation LdbcUpdate6AddPost operation containing parameter values
+     * @return Prepared UpdateQuery6 string
+     */
     public String getUpdate6Single(LdbcUpdate6AddPost operation) {
         return prepare(
-                QueryType.InteractiveUpdate6,
-                new ImmutableMap.Builder<String, String>()
-                        .put(LdbcUpdate6AddPost.POST_ID, getConverter().convertIdForInsertion(operation.getPostId()))
-                        .put(LdbcUpdate6AddPost.IMAGE_FILE, getConverter().convertString(operation.getImageFile()))
-                        .put(LdbcUpdate6AddPost.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
-                        .put(LdbcUpdate6AddPost.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
-                        .put(LdbcUpdate6AddPost.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
-                        .put(LdbcUpdate6AddPost.LANGUAGE, getConverter().convertString(operation.getLanguage()))
-                        .put(LdbcUpdate6AddPost.CONTENT, getConverter().convertString(operation.getContent()))
-                        .put(LdbcUpdate6AddPost.LENGTH, getConverter().convertInteger(operation.getLength()))
-                        .put(LdbcUpdate6AddPost.AUTHOR_PERSON_ID, getConverter().convertId(operation.getAuthorPersonId()))
-                        .put(LdbcUpdate6AddPost.FORUM_ID, getConverter().convertId(operation.getForumId()))
-                        .put(LdbcUpdate6AddPost.COUNTRY_ID, getConverter().convertId(operation.getCountryId()))
-                        .put(LdbcUpdate6AddPost.TAG_IDS, getConverter().convertLongList(operation.getTagIds()))
-                        .build()
+                QueryType.InteractiveUpdate6, getUpdate6SingleMap(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery6 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate6AddPost operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate6SingleMap(LdbcUpdate6AddPost operation) {
+        return new ImmutableMap.Builder<String, Object>()
+            .put(LdbcUpdate6AddPost.POST_ID, getConverter().convertIdForInsertion(operation.getPostId()))
+            .put(LdbcUpdate6AddPost.IMAGE_FILE, getConverter().convertString(operation.getImageFile()))
+            .put(LdbcUpdate6AddPost.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
+            .put(LdbcUpdate6AddPost.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
+            .put(LdbcUpdate6AddPost.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
+            .put(LdbcUpdate6AddPost.LANGUAGE, getConverter().convertString(operation.getLanguage()))
+            .put(LdbcUpdate6AddPost.CONTENT, getConverter().convertString(operation.getContent()))
+            .put(LdbcUpdate6AddPost.LENGTH, getConverter().convertInteger(operation.getLength()))
+            .put(LdbcUpdate6AddPost.AUTHOR_PERSON_ID, getConverter().convertId(operation.getAuthorPersonId()))
+            .put(LdbcUpdate6AddPost.FORUM_ID, getConverter().convertId(operation.getForumId()))
+            .put(LdbcUpdate6AddPost.COUNTRY_ID, getConverter().convertId(operation.getCountryId()))
+            .put(LdbcUpdate6AddPost.TAG_IDS, getConverter().convertLongList(operation.getTagIds()))
+            .build();
+    }
+
+    /**
+     * Get prepared UpdateQuery7 string
+     * This is used for single query updates.
+     * @param operation LdbcUpdate7AddComment operation containing parameter values
+     * @return Prepared UpdateQuery7 string
+     */
     public String getUpdate7Single(LdbcUpdate7AddComment operation) {
         return prepare(
-                QueryType.InteractiveUpdate7,
-                new ImmutableMap.Builder<String, String>()
-                        .put(LdbcUpdate7AddComment.COMMENT_ID, getConverter().convertIdForInsertion(operation.getCommentId()))
-                        .put(LdbcUpdate7AddComment.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
-                        .put(LdbcUpdate7AddComment.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
-                        .put(LdbcUpdate7AddComment.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
-                        .put(LdbcUpdate7AddComment.CONTENT, getConverter().convertString(operation.getContent()))
-                        .put(LdbcUpdate7AddComment.LENGTH, getConverter().convertInteger(operation.getLength()))
-                        .put(LdbcUpdate7AddComment.AUTHOR_PERSON_ID, getConverter().convertId(operation.getAuthorPersonId()))
-                        .put(LdbcUpdate7AddComment.COUNTRY_ID, getConverter().convertId(operation.getCountryId()))
-                        .put(LdbcUpdate7AddComment.REPLY_TO_POST_ID, getConverter().convertId(operation.getReplyToPostId()))
-                        .put(LdbcUpdate7AddComment.REPLY_TO_COMMENT_ID, getConverter().convertId(operation.getReplyToCommentId()))
-                        .put(LdbcUpdate7AddComment.TAG_IDS, getConverter().convertLongList(operation.getTagIds()))
-                        .build()
+                QueryType.InteractiveUpdate7, getUpdate7SingleMap(operation)
         );
     }
 
+    /**
+     * Get UpdateQuery7 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate7AddComment operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate7SingleMap(LdbcUpdate7AddComment operation) {
+        return new ImmutableMap.Builder<String, Object>()
+        .put(LdbcUpdate7AddComment.COMMENT_ID, getConverter().convertIdForInsertion(operation.getCommentId()))
+        .put(LdbcUpdate7AddComment.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
+        .put(LdbcUpdate7AddComment.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
+        .put(LdbcUpdate7AddComment.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
+        .put(LdbcUpdate7AddComment.CONTENT, getConverter().convertString(operation.getContent()))
+        .put(LdbcUpdate7AddComment.LENGTH, getConverter().convertInteger(operation.getLength()))
+        .put(LdbcUpdate7AddComment.AUTHOR_PERSON_ID, getConverter().convertId(operation.getAuthorPersonId()))
+        .put(LdbcUpdate7AddComment.COUNTRY_ID, getConverter().convertId(operation.getCountryId()))
+        .put(LdbcUpdate7AddComment.REPLY_TO_POST_ID, getConverter().convertId(operation.getReplyToPostId()))
+        .put(LdbcUpdate7AddComment.REPLY_TO_COMMENT_ID, getConverter().convertId(operation.getReplyToCommentId()))
+        .put(LdbcUpdate7AddComment.TAG_IDS, getConverter().convertLongList(operation.getTagIds()))
+        .build();
+    }
+
+    /**
+     * Get prepared UpdateQuery8 string
+     * @param operation LdbcUpdate8AddFriendship operation containing parameter values
+     * @return Prepared UpdateQuery8 string
+     */
     public String getUpdate8(LdbcUpdate8AddFriendship operation) {
         return prepare(
-                QueryType.InteractiveUpdate8,
-                ImmutableMap.of(
-                        LdbcUpdate8AddFriendship.PERSON1_ID, getConverter().convertId(operation.getPerson1Id()),
-                        LdbcUpdate8AddFriendship.PERSON2_ID, getConverter().convertId(operation.getPerson2Id()),
-                        LdbcUpdate8AddFriendship.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate())
-                )
+                QueryType.InteractiveUpdate8, getUpdate8Map(operation)
         );
     }
 
-    // Interactive updates requiring multiple queries (for some systems)
+    /**
+     * Get UpdateQuery8 Map. This map contain the name of the parameter and the value as string.
+     * @param operation LdbcUpdate8AddFriendship operation containing parameter values
+     * @return Map with parameters and values as string.
+     */
+    public Map<String, Object> getUpdate8Map(LdbcUpdate8AddFriendship operation) {
+        return ImmutableMap.of(
+            LdbcUpdate8AddFriendship.PERSON1_ID, getConverter().convertId(operation.getPerson1Id()),
+            LdbcUpdate8AddFriendship.PERSON2_ID, getConverter().convertId(operation.getPerson2Id()),
+            LdbcUpdate8AddFriendship.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate())
+        );
+    }
 
+    /**
+     * Get prepared UpdateQuery1 strings.
+     * This is used for system requiring multiple updates.
+     * @param operation UpdateQuery1 operation containing parameter values
+     * @return List of prepared UpdateQuery1 strings
+     */
     public List<String> getUpdate1Multiple(LdbcUpdate1AddPerson operation) {
         List<String> list = new ArrayList<>();
         list.add(prepare(
                 QueryType.InteractiveUpdate1AddPerson,
-                new ImmutableMap.Builder<String, String>()
+                new ImmutableMap.Builder<String, Object>()
                         .put(LdbcUpdate1AddPerson.PERSON_ID, getConverter().convertIdForInsertion(operation.getPersonId()))
                         .put(LdbcUpdate1AddPerson.PERSON_FIRST_NAME, getConverter().convertString(operation.getPersonFirstName()))
                         .put(LdbcUpdate1AddPerson.PERSON_LAST_NAME, getConverter().convertString(operation.getPersonLastName()))
@@ -563,6 +845,12 @@ public abstract class QueryStore {
         return list;
     }
 
+    /**
+     * Get prepared UpdateQuery4 strings.
+     * This is used for system requiring multiple updates.
+     * @param operation UpdateQuery4 operation containing parameter values
+     * @return List of prepared UpdateQuery4 strings
+     */
     public List<String> getUpdate4Multiple(LdbcUpdate4AddForum operation) {
         List<String> list = new ArrayList<>();
         list.add(prepare(
@@ -587,11 +875,17 @@ public abstract class QueryStore {
         return list;
     }
 
+    /**
+     * Get prepared UpdateQuery6 string
+     * This is used for system requiring multiple updates.
+     * @param operation UpdateQuery6 operation containing parameter values
+     * @return List of prepared UpdateQuery6 strings
+     */
     public List<String> getUpdate6Multiple(LdbcUpdate6AddPost operation) {
         List<String> list = new ArrayList<>();
         list.add(prepare(
                 QueryType.InteractiveUpdate6AddPost,
-                new ImmutableMap.Builder<String, String>()
+                new ImmutableMap.Builder<String, Object>()
                         .put(LdbcUpdate6AddPost.POST_ID, getConverter().convertIdForInsertion(operation.getPostId()))
                         .put(LdbcUpdate6AddPost.IMAGE_FILE, getConverter().convertString(operation.getImageFile()))
                         .put(LdbcUpdate6AddPost.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
@@ -618,56 +912,17 @@ public abstract class QueryStore {
         return list;
     }
 
-    public List<String> getUpdate6MultipleSeparatedContent(LdbcUpdate6AddPost operation) {
-        List<String> list = new ArrayList<>();
-        list.add(prepare(
-                QueryType.InteractiveUpdate6AddPost,
-                new ImmutableMap.Builder<String, String>()
-                        .put(LdbcUpdate6AddPost.POST_ID, getConverter().convertIdForInsertion(operation.getPostId()))
-                        .put(LdbcUpdate6AddPost.IMAGE_FILE, getConverter().convertString(operation.getImageFile()))
-                        .put(LdbcUpdate6AddPost.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
-                        .put(LdbcUpdate6AddPost.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
-                        .put(LdbcUpdate6AddPost.BROWSER_USED, getConverter().convertString(operation.getBrowserUsed()))
-                        .put(LdbcUpdate6AddPost.LANGUAGE, getConverter().convertString(operation.getLanguage()))
-                        .put(LdbcUpdate6AddPost.CONTENT, getConverter().convertString(operation.getContent()))
-                        .put(LdbcUpdate6AddPost.LENGTH, getConverter().convertInteger(operation.getLength()))
-                        .put(LdbcUpdate6AddPost.AUTHOR_PERSON_ID, getConverter().convertId(operation.getAuthorPersonId()))
-                        .put(LdbcUpdate6AddPost.FORUM_ID, getConverter().convertId(operation.getForumId()))
-                        .put(LdbcUpdate6AddPost.COUNTRY_ID, getConverter().convertId(operation.getCountryId()))
-                        .build()
-                )
-        );
-        for (long tagId : operation.getTagIds()) {
-            list.add(prepare(
-                    QueryType.InteractiveUpdate6AddPostTags,
-                    ImmutableMap.of(
-                            LdbcUpdate6AddPost.POST_ID, getConverter().convertIdForInsertion(operation.getPostId()),
-                            "tagId", getConverter().convertId(tagId))
-                    )
-            );
-        }
-        // Check the query, and you will see why convertId needed instead of convertIdForInsertion
-        list.add(prepare(QueryType.InteractiveUpdate6AddPostContent,
-                ImmutableMap.of(
-                        LdbcUpdate6AddPost.POST_ID, getConverter().convertId(operation.getPostId()),
-                        LdbcUpdate6AddPost.CONTENT, getConverter().convertString(operation.getContent()))
-                )
-        );
-        list.add(prepare(QueryType.InteractiveUpdate6AddPostImageFile,
-                ImmutableMap.of(
-                        LdbcUpdate6AddPost.POST_ID, getConverter().convertId(operation.getPostId()),
-                        LdbcUpdate6AddPost.IMAGE_FILE, getConverter().convertString(operation.getImageFile()))
-                )
-        );
-
-        return list;
-    }
-
+    /**
+     * Get prepared UpdateQuery7 string
+     * This is used for system requiring multiple updates.
+     * @param operation UpdateQuery7 operation containing parameter values
+     * @return List of prepared UpdateQuery7 strings
+     */
     public List<String> getUpdate7Multiple(LdbcUpdate7AddComment operation) {
         List<String> list = new ArrayList<>();
         list.add(prepare(
                 QueryType.InteractiveUpdate7AddComment,
-                new ImmutableMap.Builder<String, String>()
+                new ImmutableMap.Builder<String, Object>()
                         .put(LdbcUpdate7AddComment.COMMENT_ID, getConverter().convertIdForInsertion(operation.getCommentId()))
                         .put(LdbcUpdate7AddComment.CREATION_DATE, getConverter().convertDateTime(operation.getCreationDate()))
                         .put(LdbcUpdate7AddComment.LOCATION_IP, getConverter().convertString(operation.getLocationIp()))
@@ -691,5 +946,4 @@ public abstract class QueryStore {
         }
         return list;
     }
-
 }
